Rename module `Lang.Crucible.JVM.Numeric` to `...JVM.Translation.Numeric`.
