+++ crucible-jvm/src/Lang/Crucible/JVM/Translation.hs
@@ -99,7 +99,7 @@ import           What4.Utils.MonadST (liftST)
 
 -- crucible-jvm
 import           Lang.Crucible.JVM.Types
-import           Lang.Crucible.JVM.Numeric
+import           Lang.Crucible.JVM.Translation.Numeric
 import           Lang.Crucible.JVM.ClassRefs
 import           Lang.Crucible.JVM.Generator
 import           Lang.Crucible.JVM.Class
